Extract collection form from drawer layout wrapper.
